backend/detetction faciale/requirements-train.txt
# TensorFlow : roues Windows pour Python 3.9–3.12 (pas 3.13+ au moment des tests courants).
# Si pip ne trouve aucune version : installez Python 3.11 x64 puis recréez le venv.
tensorflow>=2.15
pandas>=2.0
numpy>=1.24,<3
scikit-learn>=1.3
matplotlib>=3.8
Pillow>=10.0